requirements.txt
fastmcp>=2.0.0
python-frontmatter>=1.1.0
pyyaml>=6.0
xxhash>=3.0.0
